+++ assignment/src/main/java/com/assignment/service/BookService.java
@@ -9,6 +9,6 @@ public interface BookService {
 	long save(Book book);
 	   Book get(Integer id);
 	   List<Book> list();
-	   void update(long id, Book book);
-	   void delete(long id);
+	   void update(Integer id, Book book);
+	   void delete(Integer id);
 }
